ccp_nl: multiplex IPC onto one netlink socket
Instead of allocating a new netlink socket per TCP connection -
this is a bad idea since the kernel by default allows only 32 netlink
sockets to be opened - have a single socket allocated on module load,
and multiplex traffic from connections to the ccp onto this.

To make this work, consolidate the headers.