time: use getnstimeofday()
Instead of using jiffies, use getnstimeofday() as the time unit.

As reported by frankc:
First, a quick background on kernel timestamps. The CCP kernel module
currently uses jiffies as its measure of time.

Jiffies is a counter incremented at a regular interval of HZ times per
second by the machine’s timing hardware. The value of HZ on our machines
is currently 250, which means that the minimum interval between two
jiffies values is 4ms. The value of HZ can be increased (another
reasonable value is 1000ms) but requires recompiling the kernel, and
still would only give a jiffies precision of 1ms. There are other
methods of obtaining higher precision timing.

Each time an ACK is received (or more precisely, each time
tcp_cong_control() is called, which may be slightly less than every
ACK), we use the following high level procedure which simply advances
the state machine and runs the action as long as we’re not supposed to
be waiting.

if (now() > next_event_time) {
     // move to next action (e.g. wait, cwnd update, report)
     // execute action
}

Next_event_time is initialized to 0. Then wait is implemented as follows
(the other actions don’t update it):

void wait(uint32_t duration) {
     next_event_time = now() + us_to_jiffies(duration)
}

1. The first issue is the granularity of jiffies. Even if duration is
100 microseconds, the next possible time in the future is 4000
microseconds.

2. The next issue is that ACKs don’t arrive consistently, and the
machine only advances with the ACK clock. Let’s say now is 0 jiffies and
we set next_event_time to 2 jiffies (so 8000 microseconds later). We
might get an ack at time 1, and then not another ack until time 4. So in
reality we had to wait ~12,000 microseconds because we didn’t happen to
get an ACK exactly at the right time.