Remove necessity of a kernel patch
This commit gets the send and receive rate sample measurements without
requiring the changes from the patched kernel.

The patched kernel provides a rate sample that reports a send interval
and receive interval for the number of new bytes delivered. Without the
patch, the kernel only exposes the max of these two numbers.

The receive interval is simple to measure without the patch; the rate
sample keeps track of the time the last ack was received, so we can
subtract the prior timestamp from the current receive time.

The send interval is calculated when packets are sent, from a
particular skb; in tcp_cong_control and tcp_in_ack_event, we cannot
access the particular skb for which the send interval is calculated.
However, we observe that the tcp rate sample code updates the variable
tp->first_tx_mstamp with the skb's scb->tx.first_tx_mstamp
(corresponding to the send time of the packet that was acked right
before this packet was sent). As long as we pick an skb from the list
where tp->first_tx_mstamp matches, the rate sample will be accurate.

In tcp_in_ack_event, we store the information for 50 SKBs forward
(the previous send timestamp, as well as the send interval as calculated
by that skb). Then in tcp cong control, after the kernel has updated the
rate sample, we iterate over the list of skbs to choose the correct one.